requirements.txt
matplotlib==3.9.2
numpy<=2.1.0
opencv_python<=4.10.0
pandas==2.2.2
#picamera2==0.3.18
Pillow==10.4.0
psutil==5.9.8
sahi==0.11.18
Shapely==2.0.6
ultralytics==8.2.13
vidstab==1.7.4
pybind11==2.13.6
cmake==3.25.0
